refactor: split example into different modules
